backend/requirements.txt
# Core dependencies
asyncio-throttle>=1.0.2
httpx>=0.25.0
python-dotenv>=1.0.0
supabase>=2.0.0

# Date/time handling
python-dateutil>=2.8.2
pytz>=2023.3

# Utilities
pydantic>=2.5.0

# Database (for SQL verification scripts)
psycopg2-binary>=2.9.9

# Development dependencies (optional)
# pytest>=7.4.0
# pytest-asyncio>=0.21.0
# black>=23.10.0
# mypy>=1.6.0
